0.4.0 preparations

+++ contactlist/views.py
@@ -1,16 +1,16 @@
 # ContactList - CTCL 2023
-# Date: June 9, 2023 - June 23, 2023
+# Date: June 9, 2023 - June 27, 2023
 # Purpose: Main application views
 
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import render
 from django.template import loader
 from django.template.defaulttags import register
+#from django.views.generic import ListView
 from datetime import datetime
 from . import lib
-from . import __version__
 from .lib import printe
-import csv
+import csv, io
 
 try:
     from .models import ContactItem
@@ -50,7 +50,10 @@ def index(request):
     allitems = tmplst
     
     navbar = lib.navbar()
-    context = {"title": "ContactList - List", "headers": columns, "data": allitems, "styling": lib.theme(request.COOKIES.get("theme"))["styling"], "navbar": navbar, "ver": __version__}
+    
+    context = lib.mkcontext(request, "ContactList - List")
+    context["headers"] = columns
+    context["data"] = allitems
     return HttpResponse(template.render(context, request))
 
 def new(request):
@@ -72,7 +75,10 @@ def new(request):
     else:
         form = ContactForm
     
-    return render(request, "new.html", {"title": "ContactList - New", "form": form, "styling": lib.theme(request.COOKIES.get("theme"))["styling"], "navbar": lib.navbar(), "ver": __version__})
+    context = lib.mkcontext(request, "ContactList - New")
+    context["form"] = form
+    return render(request, "new.html", context)
+
 
 def view(request, inid):
     template = loader.get_template("view.html")
@@ -80,12 +86,12 @@ def view(request, inid):
     data = dbitem.todict()
     # Remove database ID
     data.pop("inid")
-    
-    headers = lib.getconfig("headers")
-    
+    headers = lib.getconfig("headers")    
     navbar = lib.navbar()
-
-    context = {"title": "ContactList - View", "data": data, "headers": headers, "styling": lib.theme(request.COOKIES.get("theme"))["styling"], "navbar": navbar, "ver": __version__}
+    
+    context = lib.mkcontext(request, "ContactList - View")
+    context["headers"] = headers
+    context["data"] = data
     return HttpResponse(template.render(context, request))
 
 def edit(request, inid):
@@ -94,36 +100,64 @@ def edit(request, inid):
         form = ContactForm(request.POST, instance=data)
         if form.is_valid():
             form.save()
-            
             data = ContactItem.objects.get(pk=inid)
             data.tmod = datetime.now()
             data.save()
-            
             return HttpResponseRedirect("/")
     else:
         data = ContactItem.objects.get(pk=inid)
         form = ContactForm(initial = data.todict())
-                
-        return render(request, "edit.html", {"title": "ContactList - Edit", "form": form, "inid": inid, "styling": lib.theme(request.COOKIES.get("theme"))["styling"], "navbar": lib.navbar, "ver": __version__})
+        
+        context = lib.mkcontext(request, "ContactList - Edit")
+        context["form"] = form
+        context["inid"] = inid
+        return render(request, "edit.html", context)
 
 def delete(request, inid):
-    # TODO: Add confirmation for deleting an item
-    dbitem = ContactItem.objects.get(inid=inid)
-    dbitem.delete()
-    
-    return HttpResponseRedirect("/")
+    # The button for continuing with deletion would be a form that does not include data
+    if request.method == "POST":
+        dbitem = ContactItem.objects.get(inid=inid)
+        dbitem.delete()
+        return HttpResponseRedirect("/")
+    else:
+        return render(request, "delconfirm.html", lib.mkcontext(request, "ContactList - Delete Item"))
     
 def settings(request):
     if request.method == "POST":
         form = SettingsForm(request.POST)
         if form.is_valid():
-            
             response = HttpResponseRedirect("/")
             response.set_cookie("theme", form.cleaned_data["theme"])
             return response
     else:
         form = SettingsForm()
         
-        return render(request, "settings.html", {"title": "ContactList - Settings", "form": form, "styling": lib.theme(request.COOKIES.get("theme"))["styling"], "navbar": lib.navbar, "ver": __version__})
-            
+        context = lib.mkcontext(request, "ContactList - Settings")
+        context["form"] = SettingsForm()
+        return render(request, "settings.html", context)
+        
+def exportcsv(request):
+    allitems = [i.todict() for i in ContactItem.objects.all()]
+    fields = ContactItem.fieldnames()
+    
+    for i in allitems:
+        for x in fields:
+            # Format any datetime objects as they would be shown on the HTML table
+            if isinstance(i[x], datetime):
+                i[x] = lib.dt2fmt(i[x])
+    
+    # Create "file" in memory for DictWriter. This is done to minimize disk writes.
+    memcsv = io.StringIO()
+    writer = csv.DictWriter(memcsv, fieldnames = fields, delimiter = ",", quoting = csv.QUOTE_ALL)
+    writer.writeheader()
+    writer.writerows(allitems)
+    
+    response = HttpResponse()
+    response['Content-Disposition'] = "attachment;filename=export.csv"
+    response.write(memcsv.getvalue())
+    
+    return response
+    
+    
+
